ENH: update summary to not use skbio.io.read
